# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pickleball Club Ladder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players (search)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: exclude_team_season_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  # SEASONS
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches
      parameters:
        - name: season_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: team_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithDetails"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchWithDetails"
  /matches/{id}/score:
    post:
      operationId: submitScore
      tags: [matches]
      summary: Submit match score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoreBody"
      responses:
        "200":
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchWithDetails"
  /matches/{id}/confirm:
    post:
      operationId: confirmScore
      tags: [matches]
      summary: Confirm match score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Score confirmed, ladder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchWithDetails"
  /matches/{id}/dispute:
    post:
      operationId: disputeScore
      tags: [matches]
      summary: Dispute match score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeScoreBody"
      responses:
        "200":
          description: Dispute filed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchWithDetails"

  # NOTIFICATIONS
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current player
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsResponse"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth
...
    Player:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [player, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
...
    Season:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - startDate
        - endDate
        - isActive
        - createdAt
...
    Team:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        player1Id:
          type: string
        player2Id:
          type: string
        teamName:
          type: string
        status:
          type: string
          enum: [pending, active, inactive]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - player1Id
        - player2Id
        - teamName
        - status
        - createdAt
    TeamWithPlayers:
      allOf:
        - $ref: "#/components/schemas/Team"
        - type: object
          properties:
            player1:
              $ref: "#/components/schemas/Player"
            player2:
              $ref: "#/components/schemas/Player"
            standing:
              $ref: "#/components/schemas/LadderStanding"
            season:
              $ref: "#/components/schemas/Season"
...
    LadderStanding:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        teamId:
          type: string
        position:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        lastMatchDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - teamId
        - position
        - wins
        - losses
        - createdAt
...
    Challenge:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        challengerTeamId:
          type: string
        challengedTeamId:
          type: string
        status:
          type: string
          enum: [pending, accepted, scheduling, scheduled, completed, cancelled, disputed]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - challengerTeamId
        - challengedTeamId
        - status
        - createdAt
    ChallengeWithTeams:
      allOf:
        - $ref: "#/components/schemas/Challenge"
        - type: object
          properties:
            challengerTeam:
              $ref: "#/components/schemas/TeamWithPlayers"
            challengedTeam:
              $ref: "#/components/schemas/TeamWithPlayers"
            match:
              $ref: "#/components/schemas/MatchWithDetails"
            myTeamId:
              type: string
            overlappingSlots:
              type: array
              items:
                $ref: "#/components/schemas/TimeSlot"
            challengerAvailabilitySubmitted:
              type: boolean
            challengedAvailabilitySubmitted:
              type: boolean
...
    TimeSlot:
      type: object
      properties:
        date:
          type: string
        times:
          type: array
          items:
            type: string
      required:
        - date
        - times

    # Availability
...
    Match:
      type: object
      properties:
        id:
          type: string
        challengeId:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        courtLocation:
          type: string
        bookedByTeamId:
          type: string
        status:
          type: string
          enum: [scheduled, completed, disputed]
        createdAt:
          type: string
      required:
        - id
        - challengeId
        - scheduledDate
        - scheduledTime
        - courtLocation
        - status
        - createdAt
    MatchScore:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        gameNumber:
          type: integer
        team1Score:
          type: integer
        team2Score:
          type: integer
      required:
        - id
        - matchId
        - gameNumber
        - team1Score
        - team2Score
    MatchResult:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        winnerTeamId:
          type: string
        loserTeamId:
          type: string
        submittedByTeamId:
          type: string
        confirmedByTeamId:
          type: string
        submittedAt:
          type: string
        confirmedAt:
          type: string
        autoConfirmed:
          type: boolean
        disputeReason:
          type: string
        disputeResolved:
          type: boolean
      required:
        - id
        - matchId
        - winnerTeamId
        - loserTeamId
        - submittedByTeamId
        - submittedAt
        - autoConfirmed
        - disputeResolved
    MatchWithDetails:
      allOf:
        - $ref: "#/components/schemas/Match"
        - type: object
          properties:
            challenge:
              $ref: "#/components/schemas/ChallengeWithTeams"
            scores:
              type: array
              items:
                $ref: "#/components/schemas/MatchScore"
            result:
              $ref: "#/components/schemas/MatchResult"
    SubmitScoreBody:
      type: object
      properties:
        games:
          type: array
          items:
            type: object
            properties:
              gameNumber:
                type: integer
              team1Score:
                type: integer
              team2Score:
                type: integer
            required:
              - gameNumber
              - team1Score
              - team2Score
        winnerTeamId:
          type: string
      required:
        - games
        - winnerTeamId
    DisputeScoreBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason

    # Notification
    Notification:
      type: object
      properties:
        id:
          type: string
        playerId:
          type: string
        type:
          type: string
        message:
          type: string
        link:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - playerId
        - type
        - message
        - isRead
        - createdAt
    NotificationsResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required:
        - notifications
        - unreadCount

    # Admin
...